[Auto] refactor(opencode): move instruction discovery
