+++ src/skillsaw/lint_tree.py
@@ -6,9 +6,8 @@ from __future__ import annotations
 
 import fnmatch
 import logging
-import os
 from pathlib import Path
-from typing import Callable, Iterator, List, Optional, Set, TYPE_CHECKING, Tuple
+from typing import List, Optional, Set, TYPE_CHECKING, Tuple
 
 from .diagnostics import safe_display
 
@@ -63,6 +62,7 @@ from .formats.codex import (
     codex_inline_hooks,
     codex_inline_mcp_servers,
 )
+from .discovery.opencode import contained_instruction_glob
 from .formats import devin
 from .utils import has_apm_generated_header, read_text
 from .paths import (
@@ -73,7 +73,6 @@ from .paths import (
     safe_exists,
     safe_is_dir,
     safe_is_file,
-    safe_is_symlink,
     safe_resolve,
 )
 from .formats.promptfoo import (
@@ -156,72 +155,6 @@ _OPENCODE_CONTENT_DIRS = tuple(
 #: own terms — no rule reports the pairing itself, since which one OpenCode
 #: loads is its business rather than a defect in either file.
 _OPENCODE_CONFIG_NAMES = ("opencode.json", "opencode.jsonc")
-
-
-def _contained_instruction_glob(
-    repo_root: Path,
-    glob_base: Path,
-    pattern: str,
-    is_excluded: Callable[[Path], bool],
-) -> Iterator[Path]:
-    """Return file matches without walking outside *repo_root* or through links."""
-    if is_absolute_path(pattern) or has_parent_traversal(pattern):
-        return
-
-    parts = tuple(part for part in Path(pattern).parts if part not in ("", "."))
-    visited: Set[Tuple[Path, int]] = set()
-
-    def _descend(directory: Path, index: int) -> Iterator[Path]:
-        state = (directory, index)
-        if state in visited:
-            return
-        visited.add(state)
-
-        # Check before scandir: resolving and rejecting directory symlinks
-        # prevents a repository-controlled pattern from enumerating elsewhere.
-        if (
-            contained_resolve(directory, repo_root) != directory
-            or safe_is_symlink(directory)
-            or is_excluded(directory)
-        ):
-            return
-
-        if index == len(parts):
-            yield directory
-            return
-
-        component = parts[index]
-        if component == "**":
-            yield from _descend(directory, index + 1)
-
-        with os.scandir(directory) as scan:
-            entries = sorted(scan, key=lambda entry: entry.name)
-
-        if component == "**":
-            for entry in entries:
-                try:
-                    is_real_dir = entry.is_dir(follow_symlinks=False)
-                except OSError:
-                    continue
-                if is_real_dir:
-                    yield from _descend(directory / entry.name, index)
-            return
-
-        for entry in entries:
-            if not fnmatch.fnmatch(entry.name, component):
-                continue
-            candidate = directory / entry.name
-            if index + 1 == len(parts):
-                yield candidate
-                continue
-            try:
-                is_real_dir = entry.is_dir(follow_symlinks=False)
-            except OSError:
-                continue
-            if is_real_dir:
-                yield from _descend(candidate, index + 1)
-
-    yield from _descend(glob_base, 0)
 
 
 if TYPE_CHECKING:
@@ -799,7 +732,7 @@ def build_lint_tree(context: "RepositoryContext") -> LintTarget:
                     searched.add(search)
                     try:
                         matches = sorted(
-                            _contained_instruction_glob(
+                            contained_instruction_glob(
                                 repo_root,
                                 glob_base,
                                 pattern,
